pyproject.toml
[tool.poetry]
name = "beauty-be"
version = "0.2.0"
description = ""
authors = ["Serhii Lakodei <serhii.lakodei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.25"
ujson = "^5.9.0"
pydantic-settings = "^2.1.0"
geoalchemy2 = "^0.14.3"
python-json-logger = "^2.0.7"
asyncpg = "^0.29.0"
uvloop = "^0.19.0"
httpx = "^0.23.0"
pydantic = {extras = ["email"], version = "^2.6.1"}
python-multipart = "^0.0.9"
gunicorn = "^21.2.0"
pyjwt = "^2.8.0"
aioboto3 = "^12.3.0"
cryptography = "^42.0.5"
python-slugify = "^8.0.4"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
darglint = "^1.8.1"
flake8 = "^7.0.0"
flake8-bandit = "^4.1.1"
flake8-comprehensions = "^3.14.0"
flake8-eradicate = "^1.5.0"
flake8-isort = "^6.1.1"
flake8-spellcheck = "^0.28.0"
bandit = "^1.7.7"
isort = "^5.13.2"
mypy = "^1.8.0"
pep8-naming = "^0.13.3"
factory-boy = "^3.2.1"
pytest = "^7.1.3"
pytest-httpx = "^0.21.0"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
